campus loader, but there are differences between the directory and principals files
