+++ scuole/campuses/management/commands/bootstrapcampuses.py
@@ -49,10 +49,10 @@ class Command(BaseCommand):
             raise CommandError('A year is required.')
 
         # path to askTed file
-        askted_file = os.path.join(
-            settings.DATA_FOLDER, 'askted', 'directory.csv')
+        # askted_file = os.path.join(
+        #     settings.DATA_FOLDER, 'askted', 'directory.csv')
 
-        self.askted_data = self.load_askted_file(askted_file)
+        # self.askted_data = self.load_askted_file(askted_file)
 
         # path to FAST file where original cleaned up names are stored
         fast_file = os.path.join(
@@ -68,12 +68,12 @@ class Command(BaseCommand):
         self.shape_data = self.load_geojson_file(campus_file)
 
         # path to askTed file where principal information is stored
-        principal_file = os.path.join(
-            settings.DATA_FOLDER,
-            'askted', 'campus', 'principals.csv')
+        # principal_file = os.path.join(
+        #     settings.DATA_FOLDER,
+        #     'askted', 'campus', 'principals.csv')
 
-        self.principal_data = self.load_principal_file(
-            principal_file)
+        # self.principal_data = self.load_principal_file(
+        #     principal_file)
 
         # path to file where new, cleaned up campus names and IDs are stored
         new_campuses = os.path.join(
@@ -104,17 +104,17 @@ class Command(BaseCommand):
 
         self.make_slugs_unique()
 
-    def load_askted_file(self, file):
-        payload = {}
+    # def load_askted_file(self, file):
+    #     payload = {}
 
-        with open(file, 'r') as f:
-            reader = csv.DictReader(f)
+    #     with open(file, 'r') as f:
+    #         reader = csv.DictReader(f)
 
-            for row in reader:
-                tea_id = row['School Number'].replace("'", "")
-                payload[tea_id] = row
+    #         for row in reader:
+    #             tea_id = row['School Number'].replace("'", "")
+    #             payload[tea_id] = row
 
-        return payload
+    #     return payload
 
     def load_fast_file(self, file):
         payload = {}
@@ -139,21 +139,21 @@ class Command(BaseCommand):
 
         return payload
 
-    def load_principal_file(self, file):
-        payload = {}
+    # def load_principal_file(self, file):
+    #     payload = {}
 
-        with open(file, 'r') as f:
-            reader = csv.DictReader(f)
+    #     with open(file, 'r') as f:
+    #         reader = csv.DictReader(f)
 
-            for row in reader:
-                tea_id = row['Organization Number'].replace("'", "")
+    #         for row in reader:
+    #             tea_id = row['Organization Number'].replace("'", "")
 
-                if tea_id not in payload:
-                    payload[tea_id] = []
+    #             if tea_id not in payload:
+    #                 payload[tea_id] = []
 
-                payload[tea_id].append(row)
+    #             payload[tea_id].append(row)
 
-        return payload
+    #     return payload
 
     def load_newCampus_file(self, file):
         payload = {}
@@ -221,43 +221,19 @@ class Command(BaseCommand):
             self.stderr.write('No shape data for {}'.format(name))
             geometry = None
 
-        if campus['CAMPUS'] in self.askted_data:
-            askted_match = self.askted_data[campus['CAMPUS']]
-            phone_number = askted_match['School Phone']
-            if 'ext' in phone_number:
-                phone_number, phone_number_extension = phone_number.split(
-                    ' ext:')
-                phone_number_extension = str(phone_number_extension)
-            else:
-                phone_number_extension = ''
-            street = askted_match['School Street Address']
-            city = askted_match['School City']
-            state = askted_match['School State']
-            zip_code = askted_match['School Zip']
-            website = askted_match['School Web Page Address']
-        else:
-            self.stderr.write('No askted data for {}'.format(name))
-            phone_number = ''
-            phone_number_extension = ''
-            street = ''
-            city = ''
-            state = ''
-            zip_code = ''
-            website = ''
-
         instance, _ = Campus.objects.update_or_create(
             tea_id=campus['CAMPUS'],
             defaults={
                 'name': name,
                 'slug': slugify(name),
-                'phone_number': phone_number,
-                'phone_number_extension': phone_number_extension,
-                'website': website,
+                # 'phone_number': phone_number,
+                # 'phone_number_extension': phone_number_extension,
+                # 'website': website,
                 'charter': charter,
-                'street': street,
-                'city': city,
-                'state': state,
-                'zip_code': zip_code,
+                # 'street': street,
+                # 'city': city,
+                # 'state': state,
+                # 'zip_code': zip_code,
                 'coordinates': geometry,
                 'low_grade': low_grade,
                 'high_grade': high_grade,
@@ -267,12 +243,12 @@ class Command(BaseCommand):
             }
         )
 
-        if campus['CAMPUS'] in self.principal_data:
-            instance_principals = self.principal_data[campus['CAMPUS']]
+        # if campus['CAMPUS'] in self.principal_data:
+        #     instance_principals = self.principal_data[campus['CAMPUS']]
 
-            self.load_principals(instance, instance_principals)
-        else:
-            self.stderr.write('No principal data for {}'.format(name))
+        #     self.load_principals(instance, instance_principals)
+        # else:
+        #     self.stderr.write('No principal data for {}'.format(name))
 
     def make_slugs_unique(self):
         for district in District.objects.all():
@@ -288,36 +264,36 @@ class Command(BaseCommand):
                     campus.slug, campus.low_grade, campus.high_grade)
                 campus.save()
 
-    def load_principals(self, campus, principals):
-        for principal in principals:
-            name = '{} {}'.format(
-                principal['First Name'], principal['Last Name'])
-            name = string.capwords(name)
-            phone_number = principal['Phone']
-            fax_number = principal['Fax']
+    # def load_principals(self, campus, principals):
+    #     for principal in principals:
+    #         name = '{} {}'.format(
+    #             principal['First Name'], principal['Last Name'])
+    #         name = string.capwords(name)
+    #         phone_number = principal['Phone']
+    #         fax_number = principal['Fax']
 
-            if 'ext' in phone_number:
-                phone_number, phone_number_extension = phone_number.split(
-                    ' ext:')
-                phone_number_extension = str(phone_number_extension)
-            else:
-                phone_number_extension = ''
+    #         if 'ext' in phone_number:
+    #             phone_number, phone_number_extension = phone_number.split(
+    #                 ' ext:')
+    #             phone_number_extension = str(phone_number_extension)
+    #         else:
+    #             phone_number_extension = ''
 
-            if 'ext' in fax_number:
-                fax_number, fax_number_extension = fax_number.split(' ext:')
-                fax_number_extension = str(phone_number_extension)
-            else:
-                fax_number_extension = ''
+    #         if 'ext' in fax_number:
+    #             fax_number, fax_number_extension = fax_number.split(' ext:')
+    #             fax_number_extension = str(phone_number_extension)
+    #         else:
+    #             fax_number_extension = ''
 
-            Principal.objects.update_or_create(
-                name=name,
-                campus=campus,
-                defaults={
-                    'role': string.capwords(principal['Role']),
-                    'email': principal['Email Address'],
-                    'phone_number': phone_number,
-                    'phone_number_extension': phone_number_extension,
-                    'fax_number': fax_number,
-                    'fax_number_extension': fax_number_extension,
-                }
-            )
+    #         Principal.objects.update_or_create(
+    #             name=name,
+    #             campus=campus,
+    #             defaults={
+    #                 'role': string.capwords(principal['Role']),
+    #                 'email': principal['Email Address'],
+    #                 'phone_number': phone_number,
+    #                 'phone_number_extension': phone_number_extension,
+    #                 'fax_number': fax_number,
+    #                 'fax_number_extension': fax_number_extension,
+    #             }
+    #         )
